Update linting and testing order in check script
This commit updates the order of linting and testing commands in the check
script. The `pyright` command is now run before `pytest` to ensure type
checking is performed before running tests. The `ruff check --fix` command
is run first to automatically fix any linting issues. This change improves
the efficiency of the script by running commands in a more logical order.

Signed-off-by: John Strunk <john.strunk@gmail.com>