Add hack-check script to pre-commit hooks
This commit adds a new pre-commit hook that runs the `hack/check.sh`
script. This script is intended to perform various checks on the codebase
to ensure quality and consistency. The hook is configured to run always,
regardless of the files changed, to guarantee that the checks are performed
on every commit.

Signed-off-by: John Strunk <john.strunk@gmail.com>