Add timeformat command
Introduce a new command to get or set the current time format string used
for datetime formatting. The command accepts an optional format string
argument and validates it against strftime. If no argument is provided, it
prints the current time format. The default time format is "%Y%m%d%H%M%S".
The command is registered in the command objects dictionary and its
functionality is tested with various scenarios.

Signed-off-by: John Strunk <john.strunk@gmail.com>