Add index parameter to filename template
Added support for using the 'index' parameter in the filename template.
This allows users to include a sequential index number in their filenames.
The index is persisted across GetCommand invocations, ensuring continuous
numbering.
Updated the FileformatCommand docstring and added a test case for the new
parameter.
Also, added a module-level variable to persist the index across GetCommand
invocations.

Signed-off-by: John Strunk <john.strunk@gmail.com>